# docs/api/inventory-service-openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8084",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/inventory":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "flight-inventory-controller"
                                                               ],
                                                      "operationId":  "createInventory",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/CreateFlightInventoryRequest"
                                                                                                                              }
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK",
                                                                                    "content":  {
                                                                                                    "*/*":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/FlightInventoryResponse"
                                                                                                                           }
                                                                                                            }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                     },
                  "/api/inventory/search":  {
                                                "post":  {
                                                             "tags":  [
                                                                          "flight-inventory-controller"
                                                                      ],
                                                             "operationId":  "search",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/InventorySearchRequest"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "$ref":  "#/components/schemas/FlightInventoryResponse"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                            },
                  "/api/inventory/release":  {
                                                 "post":  {
                                                              "tags":  [
                                                                           "flight-inventory-controller"
                                                                       ],
                                                              "operationId":  "releaseHold",
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/ReleaseSeatRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                              },
                                                                                  "required":  true
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/SeatHoldResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                             },
                  "/api/inventory/hold":  {
                                              "post":  {
                                                           "tags":  [
                                                                        "flight-inventory-controller"
                                                                    ],
                                                           "operationId":  "holdSeat",
                                                           "requestBody":  {
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HoldSeatRequest"
                                                                                                                                   }
                                                                                                                    }
                                                                                           },
                                                                               "required":  true
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/SeatHoldResponse"
                                                                                                                                }
                                                                                                                 }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                          },
...
                  "/api/inventory/flight/{flightId}/reopen":  {
                                                                  "patch":  {
                                                                                "tags":  [
                                                                                             "flight-inventory-controller"
                                                                                         ],
                                                                                "operationId":  "reopenInventory",
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "flightId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "integer",
                                                                                                                      "format":  "int64"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "reason",
                                                                                                       "in":  "query",
                                                                                                       "required":  false,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK",
                                                                                                              "content":  {
                                                                                                                              "*/*":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/FlightInventoryResponse"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                                          }
                                                                                                          }
                                                                                              }
                                                                            }
                                                              },
                  "/api/inventory/flight/{flightId}/close":  {
                                                                 "patch":  {
                                                                               "tags":  [
                                                                                            "flight-inventory-controller"
                                                                                        ],
                                                                               "operationId":  "closeInventory",
                                                                               "parameters":  [
                                                                                                  {
                                                                                                      "name":  "flightId",
                                                                                                      "in":  "path",
                                                                                                      "required":  true,
                                                                                                      "schema":  {
                                                                                                                     "type":  "integer",
                                                                                                                     "format":  "int64"
                                                                                                                 }
                                                                                                  },
                                                                                                  {
                                                                                                      "name":  "reason",
                                                                                                      "in":  "query",
                                                                                                      "required":  false,
                                                                                                      "schema":  {
                                                                                                                     "type":  "string"
                                                                                                                 }
                                                                                                  }
                                                                                              ],
                                                                               "responses":  {
                                                                                                 "200":  {
                                                                                                             "description":  "OK",
                                                                                                             "content":  {
                                                                                                                             "*/*":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/FlightInventoryResponse"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                                         }
                                                                                                         }
                                                                                             }
                                                                           }
                                                             },
...
                  "/api/inventory/flight/{flightId}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "flight-inventory-controller"
                                                                                ],
                                                                       "operationId":  "getByFlightId",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "flightId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int64"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  {
                                                                                                                     "*/*":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/FlightInventoryResponse"
                                                                                                                                            }
                                                                                                                             }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
                  "/api/inventory/flight/{flightId}/history":  {
                                                                   "get":  {
                                                                               "tags":  [
                                                                                            "flight-inventory-controller"
                                                                                        ],
                                                                               "operationId":  "getHistory",
                                                                               "parameters":  [
                                                                                                  {
                                                                                                      "name":  "flightId",
                                                                                                      "in":  "path",
                                                                                                      "required":  true,
                                                                                                      "schema":  {
                                                                                                                     "type":  "integer",
                                                                                                                     "format":  "int64"
                                                                                                                 }
                                                                                                  }
                                                                                              ],
                                                                               "responses":  {
                                                                                                 "200":  {
                                                                                                             "description":  "OK",
                                                                                                             "content":  {
                                                                                                                             "*/*":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "type":  "array",
                                                                                                                                                        "items":  {
                                                                                                                                                                      "$ref":  "#/components/schemas/InventoryHistoryResponse"
                                                                                                                                                                  }
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                                         }
                                                                                                         }
                                                                                             }
                                                                           }
                                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateFlightInventoryRequest":  {
                                                                            "type":  "object",
                                                                            "properties":  {
                                                                                               "flightId":  {
                                                                                                                "type":  "integer",
                                                                                                                "format":  "int64"
                                                                                                            },
                                                                                               "aircraftId":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int64"
                                                                                                              },
                                                                                               "blockedSeats":  {
                                                                                                                    "type":  "integer",
                                                                                                                    "format":  "int32",
                                                                                                                    "minimum":  0
                                                                                                                }
                                                                                           },
                                                                            "required":  [
                                                                                             "aircraftId",
                                                                                             "flightId"
                                                                                         ]
                                                                        },
...
                                   }
                   }
}